server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.9 